Rename folder loggers -> middlewares
